Add SwhipeTableViewController superclass and inherit in both TodoItemViewController and CategoryViewController to move all swiping functionality to it and out of the Category/Item controllers.  Also move logic to set item height to this super class.
